backend/requirements.txt
fastapi
uvicorn[standard]
langchain
faiss-cpu
requests
httpx
python-dotenv
